requirements.txt
aniso8601==1.1.0
autopep8==1.2.4
click==6.6
coverage==4.3.1
cycler==0.10.0
decorator==4.0.10
eventlet==0.19.0
Flask==0.11.1
Flask-Cors==3.0.2
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
gevent==1.1.2
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.10.3
matplotlib==1.5.3
mock==2.0.0
mockredispy==2.9.3
MySQL-python==1.2.5
networkx==1.11
numpy==1.11.2
pbr==1.10.0
pep8==1.7.0
py==1.4.31
pyaml==16.12.2
pyparsing==2.1.10
pyshp==1.2.10
pytest==3.0.2
pytest-cov==2.4.0
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
redis==2.10.5
requests==2.11.1
rq==0.6.0
six==1.10.0
SQLAlchemy==1.0.15
Werkzeug==0.11.11
